Unify entry types
Simplifies the interface and allows a bunch of code to become
monomorphic.
